Fixed bugs in detection process at the top of install_apps()
- Was previously not handling repeated runs through the setup step.
- Also added a test & failure message if neither intvlm8r.py.new or intvlm8r.py found.

+++ Raspberry Pi/setup.sh
@@ -43,13 +43,18 @@ RESET="\033[0m"
 install_apps ()
 {
 
-	if [[ -f www/intvlm8r.py.new &&  -f www/intvlm8r.py ]];
+	if [[ -f www/intvlm8r.py ]];
 	then
 		echo ''
-		echo 'intvlm8r.py.new & intvlm8r.py found. Looks like this is an upgrade.'
-
-		cp -fv www/intvlm8r.py www/intvlm8r.old
-		cp -fv www/intvlm8r.py.new www/intvlm8r.py
+		if [[ -f www/intvlm8r.py.new ]];
+		then
+			echo 'intvlm8r.py.new & intvlm8r.py found. Looks like this is an upgrade.'
+			cp -fv www/intvlm8r.py www/intvlm8r.old
+			cp -fv www/intvlm8r.py.new www/intvlm8r.py
+			
+		else
+			echo "intvlm8r.py found. Looks like a repeat run through the 'start' process."
+		fi;
 		echo ''
 
 		if python3 -c 'import pkgutil; exit(not pkgutil.find_loader("libssl-dev"))';
@@ -79,7 +84,8 @@ install_apps ()
 
 		echo '====== Select Upload/Transfer options ======='
 		echo "An 'X' indicates the option is already installed"
-	else
+	elif [[ -f www/intvlm8r.py.new ]];
+	then
 		echo ''
 		echo 'intvlm8r.py.new but no intvlm8r.py found. Proceeding with a new installation.'
 		cp -fv www/intvlm8r.py.new www/intvlm8r.py
@@ -92,6 +98,12 @@ install_apps ()
 		installDropbox=1
 		installGoogle=1
 		installRsync=1
+	else
+		echo ''
+		echo "Neither intvlm8r.py.new or intvlm8r.py found. Something's amiss. Are we in the right folder?"
+		echo 'Aborting.'
+		echo ''
+		exit 1
 	fi
 	while true; do
 		echo ''
